requirements.txt
python-dotenv==1.1.1
pytest==8.4.1
pytest-dotenv==0.5.2
pytest-pythonpath==0.7.4
pytest-xdist==3.8.0
requests==2.32.5
pydantic==2.12.2
playwright==1.55.0
allure-pytest==2.15.0
allure-python-commons==2.15.0
